Refactored Marker and Map Handling in NavpointUIHelper
Re-organized some code here.
