+++ JAFDTC/UI/Base/NavpointUIHelper.cs
@@ -438,100 +438,5 @@ namespace JAFDTC.UI.Base
 
             return false;
         }
-
-        // ------------------------------------------------------------------------------------------------------------
-        //
-        // navpoint list map functions
-        //
-        // ------------------------------------------------------------------------------------------------------------
-
-        /// <summary>
-        /// create and active a map window with the specified content and observers.
-        /// </summary>
-        public static MapWindow OpenMap(IMapControlVerbHandler observer, int maxRouteLen, LLFormat coordFormat,
-                                        MapMarkerInfo.MarkerTypeMask openMask, MapMarkerInfo.MarkerTypeMask editMask,
-                                        Dictionary<string, List<INavpointInfo>> routes,
-                                        List<Threat> threatEnvironment = null,
-                                        MapImportSpec mapImport = null, MapFilterSpec mapFilter = null)
-        {
-            List<INavpointInfo> allRoutes = [ ];
-            foreach (string route in routes.Keys)
-                allRoutes.AddRange(routes[route]);
-            List<string> theaters = TheatersForNavpoints(allRoutes);
-            string theater = (theaters.Count > 0) ? theaters[0] : null;
-
-            Dictionary<string, PointOfInterest> marks = [];
-            PointOfInterestDbQuery query = new(PointOfInterestTypeMask.ANY, theater);
-            foreach (PointOfInterest poi in PointOfInterestDbase.Instance.Find(query))
-                marks[poi.UniqueID] = poi;
-
-            MapWindow mapWindow = new(true, true)
-            {
-                Theater = theater,
-                OpenMask = openMask,
-                EditMask = editMask,
-                CoordFormat = coordFormat,
-                MaxRouteLength = maxRouteLen
-            };
-            mapWindow.SetupMapContent(routes, marks, threatEnvironment, mapImport, mapFilter);
-            mapWindow.RegisterMapControlVerbObserver(observer);
-
-            mapWindow.Activate();
-
-            return mapWindow;
-        }
-
-        /// <summary>
-        /// returns the display type of the marker with the specified information. this only handles poi marker
-        /// types, reuturning null for other types
-        /// </summary>
-        public static string MarkerDisplayType(MapMarkerInfo info)
-            => info.Type switch
-            {
-                MapMarkerInfo.MarkerType.POI_SYSTEM => $"Core POI",
-                MapMarkerInfo.MarkerType.POI_USER => $"User POI",
-                MapMarkerInfo.MarkerType.POI_CAMPAIGN => $"Campaign POI",
-                _ => null
-            };
-
-        /// <summary>
-        /// returns the display name of the marker with the specified information.
-        /// </summary>
-        public static string MarkerDisplayName(MapMarkerInfo info)
-        {
-            string name = null;
-            if ((info.Type == MapMarkerInfo.MarkerType.POI_SYSTEM) ||
-                (info.Type == MapMarkerInfo.MarkerType.POI_USER) ||
-                (info.Type == MapMarkerInfo.MarkerType.POI_CAMPAIGN))
-            {
-                PointOfInterest poi = PointOfInterestDbase.Instance.Find(info.TagStr);
-                if (poi != null)
-                    name = poi.Type switch
-                    {
-                        PointOfInterestType.SYSTEM => $"POI: {poi.Name}",
-                        PointOfInterestType.USER => $"User: {poi.Name}",
-                        PointOfInterestType.CAMPAIGN => $"{poi.Campaign}: {poi.Name}",
-                        _ => throw new NotImplementedException(),
-                    };
-            }
-            return name;
-        }
-
-        /// <summary>
-        /// returns the elevation of the marker with the specified information.
-        /// </summary>
-        public static string MarkerDisplayElevation(MapMarkerInfo info, string units = "")
-        {
-            string elev = null;
-            if ((info.Type == MapMarkerInfo.MarkerType.POI_SYSTEM) ||
-                (info.Type == MapMarkerInfo.MarkerType.POI_USER) ||
-                (info.Type == MapMarkerInfo.MarkerType.POI_CAMPAIGN))
-            {
-                PointOfInterest poi = PointOfInterestDbase.Instance.Find(info.TagStr);
-                if (poi != null)
-                    elev = $"{poi.Elevation}{units}";
-            }
-            return elev;
-        }
     }
 }
